refactor: move root cmd to cmd pkg
