+++ cmd/uniflow/main.go
@@ -6,7 +6,7 @@ import (
 	"os"
 	"strings"
 
-	"github.com/siyul-park/uniflow/pkg/cmd/uniflow"
+	"github.com/siyul-park/uniflow/pkg/cmd"
 	"github.com/siyul-park/uniflow/pkg/database"
 	"github.com/siyul-park/uniflow/pkg/database/memdb"
 	"github.com/siyul-park/uniflow/pkg/database/mongodb"
@@ -86,7 +86,7 @@ func execute() error {
 	}
 	systemx.AddToScheme(st)(sc)
 
-	cmd := uniflow.NewCommand(uniflow.Config{
+	cmd := cmd.NewCommand(cmd.Config{
 		Scheme:   sc,
 		Hook:     hk,
 		Database: db,
