+++ pkg/cmd/uniflow_test.go
@@ -1,4 +1,4 @@
-package uniflow
+package cmd
 
 import (
 	"bytes"
